Checkpoint shared auto-dent progress wiring
