+++ scripts/auto-dent-stream.ts
@@ -11,6 +11,13 @@
 import type { RunResult } from './auto-dent-run.js';
 import { parseFinalRunClaim } from './auto-dent-final-claim.js';
 import { ghExec } from './auto-dent-github.js';
+import {
+  buildKaizenCycleSteps,
+  formatIssueForDisplay,
+  formatProgressStepsMarkdown,
+  upsertProgressStep,
+  type RunProgressStep,
+} from './auto-dent-progress.js';
 import { parseHookOutputs, type HookOutput } from '../src/hooks/lib/gate-signal.js';
 
 // ANSI color helpers (graceful degradation when NO_COLOR is set or not a TTY)
@@ -284,39 +291,13 @@ function pushPr(result: RunResult, prUrl: string): void {
     state: 'created',
     detail: prUrl,
     url: prUrl,
-  });
-}
-
-type ProgressStep = NonNullable<RunResult['progressSteps']>[number];
-
-function upsertProgressStep(result: RunResult, step: ProgressStep): void {
-  if (!result.progressSteps) result.progressSteps = [];
-  const existing = result.progressSteps.find((s) => s.phase === step.phase);
-  if (existing) {
-    existing.state = step.state;
-    existing.detail = step.detail;
-    existing.url = step.url;
-  } else {
-    result.progressSteps.push(step);
-  }
+  }, 'replace');
 }
 
 function updateProgressFromMarker(marker: PhaseMarker, result: RunResult): void {
   const step = progressStepFromMarker(marker);
   if (!step) return;
-  mergeProgressStep(result, step);
-}
-
-function mergeProgressStep(result: RunResult, step: ProgressStep): void {
-  if (!result.progressSteps) result.progressSteps = [];
-  const existing = result.progressSteps.find((s) => s.phase === step.phase);
-  if (!existing) {
-    result.progressSteps.push(step);
-    return;
-  }
-  existing.state = step.state || existing.state;
-  existing.detail = step.detail || existing.detail;
-  existing.url = step.url || existing.url;
+  upsertProgressStep(result, step);
 }
 
 function updateProgressFromHookOutput(output: HookOutput, result: RunResult): void {
@@ -337,14 +318,14 @@ function updateProgressFromHookOutput(output: HookOutput, result: RunResult): vo
         state: output.type === 'gate-set' ? 'pending' : 'passed',
         detail,
         url,
-      });
+      }, 'replace');
       return;
     case 'needs_pr_kaizen':
       upsertProgressStep(result, {
         phase: 'REFLECT',
         state: output.type === 'gate-set' ? 'pending' : 'completed',
         detail: output.reason,
-      });
+      }, 'replace');
       return;
     case 'needs_post_merge':
       upsertProgressStep(result, {
@@ -352,12 +333,12 @@ function updateProgressFromHookOutput(output: HookOutput, result: RunResult): vo
         state: output.type === 'gate-set' ? 'merged' : 'completed',
         detail: output.reason,
         url,
-      });
+      }, 'replace');
       return;
   }
 }
 
-function progressStepFromMarker(marker: PhaseMarker): NonNullable<RunResult['progressSteps']>[number] | null {
+function progressStepFromMarker(marker: PhaseMarker): RunProgressStep | null {
   const f = marker.fields;
   switch (marker.phase) {
     case 'PICK':
@@ -508,7 +489,7 @@ export function buildInFlightComment(
     `| **Tool calls** | ${result.toolCalls} |`,
     `| **Cost so far** | $${result.cost.toFixed(2)} |`,
     `| **Status** | ${status} |`,
-    `| **Issue worked** | ${formatIssueForComment(result.pickedIssue, result.pickedIssueTitle, repo)} |`,
+    `| **Issue worked** | ${formatIssueForDisplay(result.pickedIssue, repo, result.pickedIssueTitle)} |`,
     `| **PR generated** | ${result.prs.length > 0 ? result.prs.join(', ') : 'none yet'} |`,
     `| **Review state** | ${synthetic ? 'not applicable' : result.reviewVerdict ?? (result.prs.length > 0 ? 'pending' : 'not started')} |`,
   ];
@@ -524,86 +505,10 @@ export function buildInFlightComment(
   }
   lines.push(
     '',
-    `#### Kaizen Work Cycle`,
-    '',
-    `| Step | State | Detail | Link |`,
-    `|------|-------|--------|------|`,
+    formatProgressStepsMarkdown(result, repo),
   );
-  for (const step of buildKaizenCycleSteps(result, repo)) {
-    lines.push(`| ${step.phase} | ${step.state} | ${step.detail || '-'} | ${step.url || '-'} |`);
-  }
 
   return lines.join('\n');
-}
-
-function formatIssueForComment(issue: string | undefined, title: string | undefined, repo: string): string {
-  if (!issue) return 'unknown';
-  const url = issue.startsWith('http') ? issue : issue.replace(/^#?(\d+)$/, repo ? `https://github.com/${repo}/issues/$1` : issue);
-  return title ? `${url} — ${title}` : url;
-}
-
-const PROGRESS_PHASE_ORDER = ['PICK', 'PLAN', 'EVALUATE', 'CASE', 'IMPLEMENT', 'TEST', 'PR', 'REVIEW', 'FIX', 'MERGE', 'REFLECT', 'CLEANUP', 'STOP'];
-
-function orderedProgressSteps(steps: NonNullable<RunResult['progressSteps']>): NonNullable<RunResult['progressSteps']> {
-  return [...steps].sort((a, b) => {
-    const ai = PROGRESS_PHASE_ORDER.indexOf(a.phase);
-    const bi = PROGRESS_PHASE_ORDER.indexOf(b.phase);
-    const ao = ai === -1 ? PROGRESS_PHASE_ORDER.length : ai;
-    const bo = bi === -1 ? PROGRESS_PHASE_ORDER.length : bi;
-    return ao - bo;
-  });
-}
-
-function buildKaizenCycleSteps(result: RunResult, repo: string): NonNullable<RunResult['progressSteps']> {
-  const existing = new Map<string, NonNullable<RunResult['progressSteps']>[number]>();
-  for (const step of result.progressSteps || []) existing.set(step.phase, step);
-  const synthetic = result.pickedIssue === 'not applicable';
-  const issueDetail = formatIssueForComment(result.pickedIssue, result.pickedIssueTitle, repo);
-  const prDetail = result.prs.join(', ');
-  const defaults: NonNullable<RunResult['progressSteps']> = [
-    {
-      phase: 'PICK',
-      state: synthetic ? 'not applicable' : result.pickedIssue ? 'selected' : 'not observed',
-      detail: synthetic ? (result.pickedIssueTitle || 'synthetic task') : (result.pickedIssue ? issueDetail : ''),
-      url: synthetic || !result.pickedIssue ? undefined : formatIssueForComment(result.pickedIssue, undefined, repo),
-    },
-    { phase: 'PLAN', state: synthetic ? 'not applicable' : 'not observed', detail: synthetic ? 'synthetic test task' : '' },
-    { phase: 'EVALUATE', state: synthetic ? 'not applicable' : 'not observed', detail: synthetic ? 'synthetic test task' : '' },
-    {
-      phase: 'CASE',
-      state: synthetic ? 'not applicable' : result.cases.length > 0 ? 'created' : 'not observed',
-      detail: synthetic ? 'synthetic test task' : result.cases.join(', '),
-    },
-    { phase: 'IMPLEMENT', state: synthetic && result.prs.length > 0 ? 'done' : 'not observed', detail: synthetic && result.prs.length > 0 ? 'synthetic file committed' : '' },
-    { phase: 'TEST', state: synthetic ? 'not applicable' : 'not observed', detail: synthetic ? 'pipeline probe, not product tests' : '' },
-    { phase: 'PR', state: result.prs.length > 0 ? 'created' : 'not observed', detail: prDetail, url: result.prs[0] },
-    {
-      phase: 'REVIEW',
-      state: synthetic ? 'not applicable' : result.reviewVerdict || (result.prs.length > 0 ? 'pending' : 'not observed'),
-      detail: synthetic ? 'synthetic test task' : result.reviewVerdict || result.prs.length > 0 ? [result.reviewVerdict, result.prs[0]].filter(Boolean).join(' ') : '',
-      url: result.reviewUrls?.[0] || result.prs[0],
-    },
-    {
-      phase: 'FIX',
-      state: synthetic ? 'not applicable' : result.reviewVerdict === 'pass' || result.reviewVerdict === 'skipped' ? 'not needed' : 'not observed',
-      detail: synthetic ? 'synthetic test task' : '',
-    },
-    { phase: 'MERGE', state: result.prs.length > 0 ? 'not observed' : 'not applicable', detail: prDetail, url: result.prs[0] },
-    { phase: 'REFLECT', state: synthetic ? 'not applicable' : 'not observed', detail: synthetic ? 'synthetic test task' : '' },
-    { phase: 'CLEANUP', state: synthetic ? 'not applicable' : 'not observed', detail: synthetic ? 'synthetic test task' : '' },
-    { phase: 'STOP', state: result.stopRequested ? 'requested' : 'not requested', detail: result.stopReason || '' },
-  ];
-  for (const step of defaults) {
-    const observed = existing.get(step.phase);
-    if (!observed) continue;
-    if (synthetic && ['PLAN', 'EVALUATE', 'CASE', 'TEST', 'REVIEW', 'FIX', 'REFLECT', 'CLEANUP'].includes(step.phase)) {
-      continue;
-    }
-    step.state = observed.state || step.state;
-    step.detail = observed.detail || step.detail;
-    step.url = observed.url || step.url;
-  }
-  return orderedProgressSteps(defaults);
 }
 
 /**
